Add simple shopping cart
It is based on ReactiveCocoa bindings.

Here we have 3 main components - XIB, ViewController and Presenter.
XIB defines the UI, and hooks for ViewController (IBOutlets).
ViewController binds Presenter with these IBOutlets, and Presenter
actually implements the view logic:

* Calculate all prices when user enters any of box sizes
* Checks if name/address/zip fields are valid
* Checks if everything is valid so we can proceed

The view should respond to these accordingly:

* Update all the prices when user enters/changes any of box sizes
* Highlight name/address/zip fields by red is they are invalid
* Disable Next button and enable it only when everything is valid

ReactiveCocoa allows to do that in really declarative way, making
programming of all that stuff clean and simple. Impressing.